meta/ks-version: forbid numbers `0.{1,2,3,4,5}`
The `meta/ks-version` key was added in KS 0.6 (see [0.6 release
notes](https://github.com/kaitai-io/kaitai_struct_compiler/blob/91aeb53ebff02ebfe4c122b2bc914c7256f07266/RELEASE_NOTES.md#06-2017-02-04)),
so versions from 0.1 to 0.5 do not make any sense.

In particular, `0.1` was very likely created by mistake as
`ks-version: 0.10`, since the author of the .ksy specification did not
realize that `0.10` will be interpreted as a float and become `0.1`. As
a result, the version check that `ks-version` is supposed to activate
will not be effective, and even KSC 0.9 will try to compile the .ksy
spec, since 0.9 is obviously a higher version than 0.1.

Forbidding `0.1` as a value of `ks-version` prevents this mistake. Note
that something like `ks-version: 0.60` will not be flagged (although the
same logic applies, and the implicit interpretation as `0.6` by YAML is
probably unintentional), because `ks-version: 0.6` is valid and we can't
distinguish between the two. But we will almost certainly never use
`0.60` as a version number, so this is just a theoretical concern.