Always put `additionalProperties` after `patternProperties`
This follows the example in
https://json-schema.org/understanding-json-schema/reference/object#additionalproperties,
which is introduced by the following sentence:

> You can use `additionalProperties` with a combination of `properties`
> and `patternProperties`. (...)

I'd also argue that this order makes the most sense in terms of the
semantics of these keywords. From
https://json-schema.org/understanding-json-schema/reference/object#additionalproperties
again:

> The `additionalProperties` keyword is used to control the handling of
> extra stuff, that is, properties whose names are not listed in the
> `properties` keyword or match any of the regular expressions in the
> `patternProperties` keyword.

Based on this, I consider the trio of `properties` - `patternProperties` -
`additionalProperties` to be analogous to `if` - `else if` - `else`.